User full view

+++ server/controllers/basketController.js
@@ -5,12 +5,18 @@ class BasketController {
     async get(req, res, next) {
         try {
             const { id } = req.params
+            const { user } = req
             
             const basket = await Basket.findOne({
                 where: { id },
                 include: [{ model: BasketProduct}]
             })
-            return res.status(200).json(basket)
+            if (user.id === basket.userId) {
+                return res.status(200).json(basket)
+            }
+            else (
+                next(ApiError.forbidden(`Не свою коляску катишь, ♿️`))
+            )
         } catch (error) {
             next(ApiError.internal(error.message))
         }
@@ -19,12 +25,18 @@ class BasketController {
         try {
             const { id } = req.params
             const { productId } = req.body
+            const { user } = req
 
             const basket = await Basket.findOne({ where: { id: id } })
             const product = await Product.findOne({ where: { id: productId } })
 
-            const basket_product = BasketProduct.create({ basketId: id, productId })
-            res.json({ message: `Создана связь товар-корзина: Корзина — ${id}, Товар — ${productId}` })
+            if (user.id === basket.userId) {
+                const basket_product = BasketProduct.create({ basketId: id, productId })
+                return res.status(201).json({ message: `Создана связь товар-корзина: Корзина — ${id}, Товар — ${productId}` })
+            }
+            else (
+                next(ApiError.forbidden(`Не свою коляску катишь, ♿️`))
+            )
         } catch (error) {
             next(ApiError.internal(error.message))
         }
@@ -33,14 +45,26 @@ class BasketController {
         try {
             const { id } = req.params
             const { productId } = req.body
+            const { user } = req
 
             const basket = await Basket.findOne({ where: { id: id } })
 
-            const product = await Product.findOne({ where: { id: productId } })
+            if (user.id === basket.userId) {
+                const product = await Product.findOne({ where: { id: productId } })
+                
+                const basket_product = await BasketProduct.findOne({ where: {basketId: basket.id}})
 
-            const basket_product = await BasketProduct.destroy({ where: { productId: product.id, basketId: basket.id } })
-            console.log(`basket_product: ` + JSON.stringify(basket_product));
-            res.json(JSON.stringify(basket_product))
+                if (basket_product.productId === product.id) {
+                    const basket_product = await BasketProduct.destroy({ where: { productId: product.id, basketId: basket.id } })
+                    return res.json(JSON.stringify(basket_product))
+                }
+                else {
+                    next(ApiError.badRequest(`Нет ♿️ для удаления из корзины`))
+                }
+            }
+            else (
+                next(ApiError.forbidden(`Не свою коляску катишь, ♿️`))
+            )
         } catch (error) {
             next(ApiError.internal(error.message))
         }
